fix: standardise token + collectible details loading
