+++ apps/mobile/src/i18n/locales/en/messages.po
@@ -447,13 +447,13 @@ msgstr "Code"
 
 #: src/features/token/bitcoin/inscription-token-details.tsx:60
 #: src/features/token/bitcoin/stamp-token-details.tsx:36
-#: src/features/token/components/collectible-loading.tsx:26
+#: src/features/token/components/token-loading.tsx:33
 #: src/features/token/stacks/bns-details.tsx:33
 #: src/features/token/stacks/sip9-token-details.tsx:54
 msgid "Collectible Info"
 msgstr "Collectible Info"
 
-#: src/features/token/components/collectible-loading.tsx:29
+#: src/features/token/components/token-loading.tsx:36
 #: src/features/token/stacks/sip9-token-details.tsx:58
 msgid "Collection"
 msgstr "Collection"
@@ -600,9 +600,10 @@ msgstr "Deploying"
 msgid "Deployment failed"
 msgstr "Deployment failed"
 
-#: src/features/token/components/collectible-loading.tsx:23
 #: src/features/token/components/token-description.tsx:29
 #: src/features/token/components/token-description.tsx:43
+#: src/features/token/components/token-loading.tsx:30
+#: src/features/token/components/token-loading.tsx:78
 msgid "Description"
 msgstr "Description"
 
@@ -960,7 +961,7 @@ msgstr "Layer"
 msgid "Layer 1"
 msgstr "Layer 1"
 
-#: src/features/token/bitcoin/rune-token-details.tsx:41
+#: src/features/token/bitcoin/rune-token-details.tsx:40
 msgid "Layer 1 · Bitcoin"
 msgstr "Layer 1 · Bitcoin"
 
@@ -974,7 +975,7 @@ msgstr "Layer 1 • Bitcoin"
 msgid "Layer 2"
 msgstr "Layer 2"
 
-#: src/features/token/stacks/sip10-token-details.tsx:46
+#: src/features/token/stacks/sip10-token-details.tsx:45
 msgid "Layer 2 · Stacks"
 msgstr "Layer 2 · Stacks"
 
@@ -1008,6 +1009,7 @@ msgid "Lock failed"
 msgstr "Lock failed"
 
 #: src/features/account/components/account-stacking.tsx:52
+#: src/features/activity/utils/format-activity.ts:38
 msgid "Locked"
 msgstr "Locked"
 
@@ -1102,8 +1104,8 @@ msgstr "More options"
 #: src/features/settings/wallet-and-accounts/wallet-name-sheet.tsx:18
 #: src/features/token/bitcoin/inscription-token-details.tsx:63
 #: src/features/token/bitcoin/stamp-token-details.tsx:39
-#: src/features/token/components/collectible-loading.tsx:28
 #: src/features/token/components/token-details-table.tsx:19
+#: src/features/token/components/token-loading.tsx:35
 #: src/features/token/stacks/bns-details.tsx:35
 #: src/features/token/stacks/sip9-token-details.tsx:56
 msgid "Name"
@@ -1240,7 +1242,7 @@ msgstr "Protected Bitcoin balance:"
 
 #: src/features/token/bitcoin/inscription-token-details.tsx:67
 #: src/features/token/bitcoin/stamp-token-details.tsx:43
-#: src/features/token/components/collectible-loading.tsx:30
+#: src/features/token/components/token-loading.tsx:37
 #: src/features/token/stacks/bns-details.tsx:42
 #: src/features/token/stacks/sip9-token-details.tsx:71
 msgid "Protocol"
